+++ test/IncomeVaultRestricted.t.sol
@@ -1,82 +1,25 @@
 // SPDX-License-Identifier: MPL-2.0
-pragma solidity ^0.8.20;
+pragma solidity ^0.8.24;
 
-import "forge-std/Test.sol";
 import "./HelperContract.sol";
-import "CMTAT/interfaces/engine/IRuleEngine.sol";
-import "CMTAT/interfaces/engine/IAuthorizationEngine.sol";
-import {IncomeVault} from "../src/IncomeVault.sol";
-//import {Upgrades,} from "openzeppelin-foundry-upgrades/Upgrades.sol";
 
 /**
-* @title Test for DebtVault
+* @title Test for the restricted functions of IncomeVault
 */
-contract IncomeVaultRestrictedTest is Test, HelperContract {
+contract IncomeVaultRestrictedTest is HelperContract {
     uint256 resUint256;
-    uint8 resUint8;
     bool resBool;
-    bool resCallBool;
-    string resString;
-    uint8 CODE_NONEXISTENT = 255;
-
-    // ADMIN balance payment
-    uint256 tokenBalance = 5000;
 
     // Arrange
     function setUp() public {
-        // Deploy CMTAT
-            CMTAT_CONTRACT = new CMTAT_STANDALONE(
-            ZERO_ADDRESS,
-            CMTAT_ADMIN,
-            IAuthorizationEngine(address(0)),
-            "CMTA Token",
-            "CMTAT",
-            DECIMALS,
-            "CMTAT_ISIN",
-            "https://cmta.ch",
-            IRuleEngine(address(0)),
-            "CMTAT_info",
-            FLAG
-        );
-
-        // Token payment
-        tokenPayment = new CMTAT_STANDALONE(
-            ZERO_ADDRESS,
-            TOKEN_PAYMENT_ADMIN,
-            IAuthorizationEngine(address(0)),
-            "CMTA Token",
-            "CMTAT",
-            DECIMALS,
-            "CMTAT_ISIN",
-            "https://cmta.ch",
-            IRuleEngine(address(0)),
-            "CMTAT_info",
-            FLAG
-        );
-        Options memory opts;
-        opts.constructorData = abi.encode(ZERO_ADDRESS);
-        address proxy = Upgrades.deployTransparentProxy(
-            "IncomeVault.sol",
-            DEFAULT_ADMIN_ADDRESS,
-            abi.encodeCall(IncomeVault.initialize, ( DEFAULT_ADMIN_ADDRESS,
-            tokenPayment,
-            ICMTATSnapshot(address(CMTAT_CONTRACT)),
-            IRuleEngine(ZERO_ADDRESS),
-            IAuthorizationEngine(ZERO_ADDRESS),
-            TIME_LIMIT_TO_WITHDRAW)),
-            opts
-        );
-        debtVault = IncomeVault(proxy);
-        vm.prank(TOKEN_PAYMENT_ADMIN);
-        tokenPayment.mint(DEFAULT_ADMIN_ADDRESS, tokenBalance);
-
+        _deployContracts();
     }
 
     function testDepositRoleCanPerformDeposit() public {
         uint256 time = 200;
         // Allowance
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        tokenPayment.approve(address(debtVault), defaultDepositAmount);
+        tokenPayment.approve(address(incomeVault), defaultDepositAmount);
         // Act
         vm.prank(DEFAULT_ADMIN_ADDRESS);
         //Event
@@ -86,55 +29,39 @@ contract IncomeVaultRestrictedTest is Test, HelperContract {
             DEFAULT_ADMIN_ADDRESS,
             defaultDepositAmount
         );
-        debtVault.deposit(time, defaultDepositAmount);
+        incomeVault.deposit(time, defaultDepositAmount);
         // Assert
-        resUint256 = debtVault.segregatedDividend(time);
-        assertEq(resUint256, defaultDepositAmount); 
+        resUint256 = incomeVault.segregatedDividend(time);
+        assertEq(resUint256, defaultDepositAmount);
     }
 
-
-    function _performOnlyDeposit() internal {
-        // Allowance
+    function testCannotDepositZeroAmount() public {
+        vm.expectRevert(
+        abi.encodeWithSelector(IncomeVault_NoAmountSend.selector));
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        tokenPayment.approve(address(debtVault), defaultDepositAmount);
-        // Act
-        vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.deposit(defaultSnapshotTime, defaultDepositAmount);
-    }
-
-    function _performDeposit() internal {
-        _performOnlyDeposit();
-        // Configure snapshot
-
-        vm.prank(CMTAT_ADMIN);
-        CMTAT_CONTRACT.scheduleSnapshot(defaultSnapshotTime);
-        
-        // Mint token for Address 1
-        vm.prank(CMTAT_ADMIN);
-        CMTAT_CONTRACT.mint(ADDRESS1, ADDRESS1_INITIAL_AMOUNT);
+        incomeVault.deposit(200, 0);
     }
 
     function testAdminCanWithdrawAll() public {
         // Arrange
-        // Deposit
         uint256 snapshotTime1 = block.timestamp + 50;
-        uint256 snapshotTime2 = block.timestamp + 50;
+        uint256 snapshotTime2 = block.timestamp + 100;
         uint256 depositAmount1 = 2000;
         uint256 depositAmount2 = 3000;
-        uint256 ALLOWANCE_NEEDED = 2000 + 3000;
+        uint256 ALLOWANCE_NEEDED = depositAmount1 + depositAmount2;
         // Allowance
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        tokenPayment.approve(address(debtVault), ALLOWANCE_NEEDED);
+        tokenPayment.approve(address(incomeVault), ALLOWANCE_NEEDED);
         // Deposit 1
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.deposit(snapshotTime1, depositAmount1);
+        incomeVault.deposit(snapshotTime1, depositAmount1);
         // Deposit 2
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.deposit(snapshotTime2, depositAmount2);
-        
+        incomeVault.deposit(snapshotTime2, depositAmount2);
+
         // Withdraw
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.withdrawAll(ALLOWANCE_NEEDED, ADDRESS2);
+        incomeVault.withdrawAll(ALLOWANCE_NEEDED, ADDRESS2);
 
         // Assert
         assertEq(tokenPayment.balanceOf(ADDRESS2), ALLOWANCE_NEEDED);
@@ -142,35 +69,63 @@ contract IncomeVaultRestrictedTest is Test, HelperContract {
 
     function testAdminCanWithdrawSpecificTime() public {
         // Arrange
-        // Deposit
         uint256 snapshotTime1 = block.timestamp + 50;
-        uint256 snapshotTime2 = block.timestamp + 50;
+        uint256 snapshotTime2 = block.timestamp + 100;
         uint256 depositAmount1 = 2000;
         uint256 depositAmount2 = 3000;
-        uint256 ALLOWANCE_NEEDED = 2000 + 3000;
+        uint256 ALLOWANCE_NEEDED = depositAmount1 + depositAmount2;
         // Allowance
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        tokenPayment.approve(address(debtVault), ALLOWANCE_NEEDED);
+        tokenPayment.approve(address(incomeVault), ALLOWANCE_NEEDED);
         // Deposit 1
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.deposit(snapshotTime1, depositAmount1);
+        incomeVault.deposit(snapshotTime1, depositAmount1);
         // Deposit 2
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.deposit(snapshotTime2, depositAmount2);
-        
+        incomeVault.deposit(snapshotTime2, depositAmount2);
+
         // Withdraw
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.withdraw(snapshotTime1, depositAmount1, ADDRESS2);
+        incomeVault.withdraw(snapshotTime1, depositAmount1, ADDRESS2);
 
         // Assert
-        assertEq(tokenPayment.balanceOf(ADDRESS2),depositAmount1);
+        assertEq(tokenPayment.balanceOf(ADDRESS2), depositAmount1);
+        assertEq(incomeVault.segregatedDividend(snapshotTime1), 0);
+        assertEq(incomeVault.segregatedDividend(snapshotTime2), depositAmount2);
+    }
+
+    function testCannotWithdrawMoreThanDepositedForATime() public {
+        uint256 time = block.timestamp + 50;
+        _performOnlyDeposit(time, defaultDepositAmount);
+        vm.expectRevert(
+        abi.encodeWithSelector(IncomeVault_NotEnoughAmount.selector));
+        vm.prank(DEFAULT_ADMIN_ADDRESS);
+        incomeVault.withdraw(time, defaultDepositAmount + 1, ADDRESS2);
+    }
+
+    function testDistributeRoleCanDistributeDividend() public {
+        // Arrange
+        _performDeposit();
+        vm.warp(defaultSnapshotTime + 50);
+        vm.prank(DEFAULT_ADMIN_ADDRESS);
+        incomeVault.setStatusClaim(defaultSnapshotTime, true);
+
+        // Act
+        address[] memory addresses = new address[](1);
+        addresses[0] = ADDRESS1;
+        vm.prank(DEFAULT_ADMIN_ADDRESS);
+        incomeVault.distributeDividend(addresses, defaultSnapshotTime);
+
+        // Assert
+        assertEq(tokenPayment.balanceOf(ADDRESS1), defaultDepositAmount);
+        assertEq(incomeVault.claimedDividend(ADDRESS1, defaultSnapshotTime), true);
     }
 
     function testCanAdminSetStatusClaim() public {
         uint256 time = 122;
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.setStatusClaim(time, true);
-        resBool = debtVault.segregatedClaim(time);
+        incomeVault.setStatusClaim(time, true);
+        resBool = incomeVault.segregatedClaim(time);
         assertEq(resBool, true);
     }
 
@@ -178,12 +133,12 @@ contract IncomeVaultRestrictedTest is Test, HelperContract {
         // Arrange
         uint256 time = 122;
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.setStatusClaim(time, true);
+        incomeVault.setStatusClaim(time, true);
         // Act
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.setStatusClaim(time, false);
+        incomeVault.setStatusClaim(time, false);
         // Assert
-        resBool =  debtVault.segregatedClaim(time);
+        resBool = incomeVault.segregatedClaim(time);
         assertEq(resBool, false);
     }
 
@@ -191,53 +146,67 @@ contract IncomeVaultRestrictedTest is Test, HelperContract {
         // Act
         uint256 time = 122;
         vm.prank(DEFAULT_ADMIN_ADDRESS);
-        debtVault.setTimeLimitToWithdraw(time);
+        incomeVault.setTimeLimitToWithdraw(time);
         // Assert
-        resUint256 =  debtVault.timeLimitToWithdraw();
-        assertEq(resUint256,time);
+        resUint256 = incomeVault.timeLimitToWithdraw();
+        assertEq(resUint256, time);
     }
 
     /****** Attacker */
     function testCannotAttackerSetStatusClaim() public {
         vm.expectRevert(
-        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_OPERATOR_ROLE));  
+        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_OPERATOR_ROLE));
         vm.prank(ATTACKER);
-        debtVault.setStatusClaim(122, true);
+        incomeVault.setStatusClaim(122, true);
     }
 
     function testCannotAttackerSetTimeLimitToWithdraw() public {
         vm.expectRevert(
-        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_OPERATOR_ROLE));  
+        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_OPERATOR_ROLE));
         vm.prank(ATTACKER);
-        debtVault.setTimeLimitToWithdraw(122);
+        incomeVault.setTimeLimitToWithdraw(122);
     }
 
     function testCannotAttackerDistributeDividend() public {
         vm.expectRevert(
-        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_DISTRIBUTE_ROLE));  
+        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_DISTRIBUTE_ROLE));
         vm.prank(ATTACKER);
         address[] memory addresses = new address[](0);
-        debtVault.distributeDividend(addresses, 12);
+        incomeVault.distributeDividend(addresses, 12);
     }
 
     function testCannotAttackerWithdrawAll() public {
         vm.expectRevert(
-        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_WITHDRAW_ROLE));  
+        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_WITHDRAW_ROLE));
         vm.prank(ATTACKER);
-        debtVault.withdrawAll(12,ADDRESS2 );
+        incomeVault.withdrawAll(12, ADDRESS2);
     }
 
     function testCannotAttackerWithdraw() public {
         vm.expectRevert(
-        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_WITHDRAW_ROLE));  
+        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_WITHDRAW_ROLE));
         vm.prank(ATTACKER);
-        debtVault.withdraw(12, 12, ADDRESS2 );
+        incomeVault.withdraw(12, 12, ADDRESS2);
     }
 
     function testCannotAttackerPerformDeposit() public {
         vm.expectRevert(
-        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_DEPOSIT_ROLE));  
+        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, INCOME_VAULT_DEPOSIT_ROLE));
         vm.prank(ATTACKER);
-        debtVault.deposit(12, 12);
+        incomeVault.deposit(12, 12);
+    }
+
+    function testCannotAttackerSetRuleEngine() public {
+        vm.expectRevert(
+        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, bytes32(0)));
+        vm.prank(ATTACKER);
+        incomeVault.setRuleEngine(IRuleEngine(ADDRESS3));
+    }
+
+    function testCannotAttackerPause() public {
+        vm.expectRevert(
+        abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, incomeVault.PAUSER_ROLE()));
+        vm.prank(ATTACKER);
+        incomeVault.pause();
     }
 }